+++ src/main/java/tech/endorsed/signport/world/Anchor.java
@@ -1,6 +1,6 @@
 package tech.endorsed.signport.world;
 
-import net.minecraft.util.math.BlockPos;
+import net.minecraft.core.BlockPos;
 
 public class Anchor {
     public String name;